refactor: simplify — share the temp-git-repo fixture across terminal-workspace tests
endpoint.test.ts and serveFsGit.test.ts each carried a near-identical mkTempRepo(); extract one makeTempRepo() into gitRepo.testlib.ts (the /be simplify pass). No behavior change; terminal-workspace unit tests stay green (27).
