+++ packages/terminal-workspace/src/endpoint.test.ts
@@ -1,4 +1,3 @@
-import { execFileSync } from "node:child_process";
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
@@ -6,30 +5,14 @@ import { ORPCError } from "@orpc/server";
 import pino from "pino";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { createTerminalWorkspaceEndpoint } from "./endpoint.ts";
+import { makeTempRepo } from "./gitRepo.testlib.ts";
 
 const log = pino({ level: "silent" });
-
-/** A temp git repo with one committed file, an uncommitted edit to it, and an
- *  untracked file — so listAll/getStatus/getDiff all have something to report. */
-function mkTempRepo(): string {
-  const repo = fs.mkdtempSync(path.join(os.tmpdir(), "kolu-tw-endpoint-"));
-  const git = (...args: string[]) =>
-    execFileSync("git", args, { cwd: repo, stdio: "pipe" });
-  git("init", "-q");
-  git("config", "user.email", "test@example.com");
-  git("config", "user.name", "Test");
-  fs.writeFileSync(path.join(repo, "a.txt"), "one\n");
-  git("add", "a.txt");
-  git("commit", "-q", "-m", "init");
-  fs.writeFileSync(path.join(repo, "a.txt"), "one\ntwo\n");
-  fs.writeFileSync(path.join(repo, "untracked.txt"), "x\n");
-  return repo;
-}
 
 describe("createTerminalWorkspaceEndpoint", () => {
   let repo: string;
   beforeEach(() => {
-    repo = mkTempRepo();
+    repo = makeTempRepo();
   });
   afterEach(() => fs.rmSync(repo, { recursive: true, force: true }));
